Replace WOPublic symlink with a submodule
Given that login-tool is soon to be released as an open source product I
don't want possible future refactorings in WOPublic to break its
buildability.

Use a Git submodule to peg WOPublic to a specific commit inside the
login-tool root directory.

The cost is that in order to propagate WOPublic updates I'll have to
explicitly update the submodule, but the corresponding benefit is that
unexpected updates can never occur (and never break a previously working
build).

Another benefit is that developers can clone the login-tool superproject
and automatically pull down one of its dependencies without having to
worry about performing multiple clones and ensuring that the project
folders are in the correct relative locations.

In a future commit will look at making a similar change to the Wincent
Build Tools (currently referenced via relative path).

Signed-off-by: Wincent Colaiuta <win@wincent.com>